Started on support for telemetry
