+++ types.go
@@ -1,7 +1,6 @@
-package lib
+package notificationhubs
 
 import (
-	"strings"
 	"time"
 )
 
@@ -19,31 +18,14 @@ const (
 	paramSaasKeyValue = "SharedAccessKey="
 
 	// Http methods
-	postMethod = "POST"
-	getMethod  = "GET"
-	putMethod  = "PUT"
-)
+	deleteMethod = "DELETE"
+	getMethod    = "GET"
+	postMethod   = "POST"
+	putMethod    = "PUT"
 
-// Public constants
-const (
-	// Template notification
-	Template NotificationFormat = "template"
-	// GcmFormat (ted) notification
-	GcmFormat NotificationFormat = "gcm"
-	// AppleFormat (ted) notification
-	AppleFormat NotificationFormat = "apple"
-	// BaiduFormat (ted) notification
-	BaiduFormat NotificationFormat = "baidu"
-	// KindleFormat (ted) notification
-	KindleFormat NotificationFormat = "adm"
-	// WindowsFormat (ted) notification
-	WindowsFormat NotificationFormat = "windows"
-	// WindowsPhoneFormat (ted) notification
-	WindowsPhoneFormat NotificationFormat = "windowsphone"
-
-	// AppleRegTemplate is the XML string for registering an iOS device
+	// appleRegXMLString is the XML string for registering an iOS device
 	// Replace {{Tags}} and {{DeviceID}} with the correct values
-	AppleRegTemplate string = `<?xml version="1.0" encoding="utf-8"?>
+	appleRegXMLString string = `<?xml version="1.0" encoding="utf-8"?>
 <entry xmlns="http://www.w3.org/2005/Atom">
   <content type="application/xml">
     <AppleRegistrationDescription xmlns:i="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://schemas.microsoft.com/netservices/2010/10/servicebus/connect">
@@ -53,15 +35,41 @@ const (
   </content>
 </entry>`
 
-	// GcmRegTemplate is the XML string for registering an iOS device
+	// appleTemplateRegXMLString is the XML string for registering an iOS device
+	// Replace {{Tags}}, {{DeviceID}} and {{Template}} with the correct values
+	appleTemplateRegXMLString string = `<?xml version="1.0" encoding="utf-8"?>
+<entry xmlns="http://www.w3.org/2005/Atom">
+  <content type="application/xml">
+    <AppleTemplateRegistrationDescription xmlns:i="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://schemas.microsoft.com/netservices/2010/10/servicebus/connect">
+      <Tags>{{Tags}}</Tags>
+      <DeviceToken>{{DeviceID}}</DeviceToken>
+			<BodyTemplate><![CDATA[{{Template}}]]></BodyTemplate>
+    </AppleTemplateRegistrationDescription>
+  </content>
+</entry>`
+
+	// gcmRegXMLString is the XML string for registering an iOS device
 	// Replace {{Tags}} and {{DeviceID}} with the correct values
-	GcmRegTemplate string = `<?xml version="1.0" encoding="utf-8"?>
+	gcmRegXMLString string = `<?xml version="1.0" encoding="utf-8"?>
 <entry xmlns="http://www.w3.org/2005/Atom">
   <content type="application/xml">
     <GcmRegistrationDescription xmlns:i="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://schemas.microsoft.com/netservices/2010/10/servicebus/connect">
       <Tags>{{Tags}}</Tags>
       <GcmRegistrationId>{{DeviceID}}</GcmRegistrationId>
     </GcmRegistrationDescription>
+  </content>
+</entry>`
+
+	// gcmRegTemplateXMLString is the XML string for registering an Android device
+	// Replace {{Tags}}, {{DeviceID}} and {{Template}} with the correct values
+	gcmTemplateRegXMLString string = `<?xml version="1.0" encoding="utf-8"?>
+<entry xmlns="http://www.w3.org/2005/Atom">
+  <content type="application/xml">
+    <GcmTemplateRegistrationDescription xmlns:i="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://schemas.microsoft.com/netservices/2010/10/servicebus/connect">
+      <Tags>{{Tags}}</Tags>
+      <GcmRegistrationId>{{DeviceID}}</GcmRegistrationId>
+			<BodyTemplate><![CDATA[{{Template}}]]></BodyTemplate>
+    </GcmTemplateRegistrationDescription>
   </content>
 </entry>`
 )
@@ -116,40 +124,45 @@ type (
 		Tags                 []string  `xml:"-"                 json:"tags,omitempty"`
 		TagsString           *string   `xml:"Tags"              json:"-"`
 	}
+
+	// NotificationDetails is the detailed information about a sent or scheduled message
+	NotificationDetails struct {
+		ID                string               `xml:"NotificationId"`
+		State             NotificationState    `xml:"State"`
+		EnqueueTime       string               `xml:"EnqueueTime"`
+		StartTime         string               `xml:"StartTime"`
+		EndTime           string               `xml:"EndTime"`
+		Body              string               `xml:"NotificationBody"`
+		TargetPlatforms   string               `xml:"TargetPlatforms"`
+		ApnsOutcomeCounts NotificationOutcomes `xml:"ApnsOutcomeCounts"`
+		GcmOutcomeCounts  NotificationOutcomes `xml:"GcmOutcomeCounts"`
+	}
+
+	// NotificationTelemetry is the id of a sent or scheduled message
+	NotificationTelemetry struct {
+		NotificationMessageID string `json:"notificationMessageId"`
+	}
+
+	// NotificationOutcomes array of outcomes
+	NotificationOutcomes struct {
+		Outcomes []NotificationOutcome `xml:"Outcome"`
+	}
+
+	// NotificationOutcome name value pair for statistics
+	NotificationOutcome struct {
+		Name  NotificationOutcomeName `xml:"Name"`
+		Count int                     `xml:"Count"`
+	}
+
+	// NotificationState is the state of the notification
+	NotificationState string
+
+	// NotificationFormat is the format of a notification
+	NotificationFormat string
+
+	// NotificationOutcomeName is a possible outcome of a notification
+	NotificationOutcomeName string
+
+	// TargetPlatform is the specific platform
+	TargetPlatform string
 )
-
-// Normalize normalizes all devices in the feed
-func (r *Registrations) normalize() {
-	for _, entry := range r.Entries {
-		if entry.RegistrationContent != nil {
-			entry.RegistrationContent.normalize()
-		}
-	}
-}
-
-// Normalize normalizes the different devices
-func (r *RegistrationContent) normalize() {
-	if r.AppleRegistrationDescription != nil {
-		r.Format = AppleFormat
-		r.RegistratedDevice = r.AppleRegistrationDescription
-		r.AppleRegistrationDescription = nil
-		r.RegistratedDevice.DeviceID = *r.RegistratedDevice.DeviceToken
-		r.RegistratedDevice.DeviceToken = nil
-	} else if r.GcmRegistrationDescription != nil {
-		r.Format = GcmFormat
-		r.RegistratedDevice = r.GcmRegistrationDescription
-		r.GcmRegistrationDescription = nil
-		r.RegistratedDevice.DeviceID = *r.RegistratedDevice.GcmRegistrationID
-		r.RegistratedDevice.GcmRegistrationID = nil
-	}
-	expirationTime, err := time.Parse("2006-01-02T15:04:05.000Z", *r.RegistratedDevice.ExpirationTimeString)
-	if err != nil { // The API uses more than one date format unfortunately :(
-		expirationTime, _ = time.Parse("2006-01-02T15:04:05.000", *r.RegistratedDevice.ExpirationTimeString)
-	}
-	r.RegistratedDevice.ExpirationTime = expirationTime
-	r.RegistratedDevice.ExpirationTimeString = nil
-	if r.RegistratedDevice != nil {
-		r.RegistratedDevice.Tags = strings.Split(*r.RegistratedDevice.TagsString, ",")
-		r.RegistratedDevice.TagsString = nil
-	}
-}
